pyproject.toml
[project]
name = "autonovel"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "python-dotenv>=1.2.2",
    "openai>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
]
